server/src/main/java/com/farmcollector/service/CropService.java
package com.farmcollector.service;

import com.farmcollector.model.Crop;
import com.farmcollector.repository.CropRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.List;
import java.util.Optional;

@Service
public class CropService {

    @Autowired
    private CropRepository cropRepository;

    public List<Crop> getAllCrops() {
        return cropRepository.findAll();
    }

    public Optional<Crop> getCropById(Long cropId) {
        return cropRepository.findById(cropId);
    }

    public Crop saveCrop(Crop crop) {
        return cropRepository.save(crop);
    }

    public void deleteCrop(Long cropId) {
        cropRepository.deleteById(cropId);
    }
}
